+++ src/blocks/results/loading/cards-list/card/__img/card__img.css
@@ -1,6 +1,7 @@
 
 .card__img {
   width: 100%;
+  height: 272px;
   border-radius: 16px 16px 0 0;
   display: block;
 }
