+++ src/js/constants/constants.js
@@ -1,0 +1,28 @@
+export default {
+mainPopup: document.querySelector('.popup'),
+popupContent: document.querySelector('.popup__content'),
+popupSignin: document.querySelector('#popup-signin'),
+popupSignup: document.querySelector('#popup-signup'),
+popupOk: document.querySelector('#popup-ok'),
+
+
+RESULTS_CONTAINER: document.querySelector('.results'),
+NEWSCARDS_CONTAINER: document.querySelector('.loading__cards'),
+CARDS_LIST: document.querySelector('.cards-list'),
+SHOWMORE_BUTTON: document.querySelector('.button__more'),
+SEARCH_FORM: document.querySelector('.search'),
+
+
+
+containerHeader: document.querySelector('.container__header'),
+headerLoginWhite: document.querySelector('#header-login_white'),
+headerLoginBlack: document.querySelector('#header-login_black'),
+headerLogout: document.querySelector('#header-logout'),
+
+
+
+KEYWORDS_ALL: document.getElementById("keywords__all"),
+KEYWORD_1: document.getElementById("keyword__1"),
+KEYWORD_2: document.getElementById("keyword__2"),
+KEYWORD_OTHERS: document.getElementById("keywords__others"),
+};
