+++ src/js/save.js
@@ -1,24 +1,13 @@
 import '../styles/save.css';
 
+import MainApi from './api/MainApi';
+import Header from './components/Header';
+import NewsCard from './components/NewsCard';
+import NewsCardList from './components/NewsCardList';
+import KeyWordsCounter from './components/KeyWordsCounter'
+import constants from './constants/constants';
+const { popupSignin, RESULTS_CONTAINER, NEWSCARDS_CONTAINER, CARDS_LIST, SHOWMORE_BUTTON, KEYWORDS_ALL, KEYWORD_1, KEYWORD_2, KEYWORD_OTHERS } = constants;
 
-const mainPopup = document.querySelector('.popup');
-const popupContent = document.querySelector('.popup__content');
-const popupSignin = document.querySelector('#popup-signin');
-const popupSignup = document.querySelector('#popup-signup');
-const popupOk = document.querySelector('#popup-ok');
-// const buttonAuth = document.querySelector('.button__auth');
-const buttonClosePopup = document.querySelector('.popup__close');
-// const popupOption = document.querySelector('.popup__option');
-
-
-const RESULTS_CONTAINER = document.querySelector('.results');
-const NEWSCARDS_CONTAINER = document.querySelector('.loading__cards');
-const CARDS_LIST = document.querySelector('.cards-list__saved');
-const SHOWMORE_BUTTON = document.querySelector('.button__more');
-const SEARCH_FORM = document.querySelector('.search');
-const SEARCH_BUTTON = document.querySelector('.search__field');
-// const FORM = document.querySelector('.popup__form')
-const UP_FORM = document.forms.signup;
 const MONTHS = [
   'Января', 'Февраля', 'Марта', 'Апреля', 'Мая', 'Июня', 'Июля',
   'Августа', 'Сентября', 'Октября', 'Ноября', 'Декабря',
@@ -49,835 +38,24 @@ const MONTHS = [
   };
 
 
-class BaseComponent {
-  constructor() {
-    this._listeners = [];
-  }
-
-  _setHandlers (handlers) {
-    handlers.forEach((handler) => {
-      this._setAddEventListener(handler);
-    });
-  }
-
-  _setAddEventListener ({ element, event, callback }) {
-    element.addEventListener(event, callback);
-    this._listeners.push({ element, event, callback });
-  }
-
-  removeListeners () {
-    this._listeners.forEach(({ element, event, callback }) => {
-      element.removeEventListener(event, callback);
-    });
-  }
-}
-
 
 
 //основной функционал
 
-class NewsApi {
-  constructor() {
-    this._deployURL = 'https://praktikum.tk/news/v2/everything';
-    this._localhostURL = 'https://newsapi.org/v2/everything';
-    this._apiKey = '9b730c26c4c34bcf9df75180873381d6';
-    this._pageSize = 100;
-  }
-
-  getNews(keyword, from, to) {
-    return fetch(`${this._localhostURL}?q=${encodeURI(keyword)}&apiKey=${this._apiKey}&language=ru&from=${from}&to=${to}&pageSize=${this._pageSize}`,
-
-    {method: 'GET'})
-      .then((res) => {
-        if (res.ok) {
-          return res.json();
-        }
-        return Promise.reject(new Error(`Код ошибки: ${res.statusCode}, Сообщение: ${res.statusText}`));
-      })
-      .then((res) => {
-        if (res.totalResults > 0) {
-          return res;
-        }
-        return Promise.reject('По запросу ничего найдено');
-      })
-      .catch((err) => Promise.reject(err));
-  }
-
-
-}
-
-const newsApi = new NewsApi();
-
-
-class MainApi {
-  constructor() {
-    this._mainURL = 'http://api.news-bord.students.nomoreparties.co';
-    this.isLoggedIn = false;
-  }
-
-  signup({email, password, name}) {
-    return fetch(`${this._mainURL}/signup`, {
-      method: 'POST',
-      credentials: 'include',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        email,
-        password,
-        name,
-      }),
-    })
-    .catch(err => {
-      console.log(err);
-    })
-    }
-
-
-  signin({email, password}) {
-    return fetch(`${this._mainURL}/signin`, {
-      method: 'POST',
-      credentials: 'include',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        email,
-        password,
-      }),
-    })
-    .then((res) => {
-      if (res.status !== 200) {
-        throw (err => {
-          console.log(err);
-        })
-      }
-      return res.json();
-    })
-    .catch(err => {
-      console.log(err);
-    })
-
-    }
-
-  getUserData() {
-    return fetch(`${this._mainURL}/users/me`, {
-      method: 'GET',
-      // credentials: 'include',
-      headers: {
-        authorization: `Bearer ${localStorage.getItem('token')}`,
-      },
-    })
-    .then((res) => {
-      if (res.status !== 200) {
-        throw (err => {
-          console.log(err);
-        })
-      }
-      return res.json();
-    })
-    .catch(err => {
-      console.log(err);
-    })
-  }
-
-
-  // getArticles() {
-  //   return this.makeFetch(`articles`);
-  // }
-
-  createArticle(article) {
-    return fetch(`${this._mainURL}/articles`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        authorization: `Bearer ${localStorage.getItem('token')}`,
-      },
-      body: JSON.stringify({
-        // keyword: article.keyword,
-        keyword: article.keyword,
-        link: article.link,
-        image: article.image,
-        title: article.title,
-        text: article.text,
-        date: article.date,
-        source: article.source,
-      }),
-    })
-    .then((res) => {
-      return res.json();
-    })
-    .catch(err => {
-      console.log(err);
-    })
-  }
-
-  removeArticle(articleId) {
-    return fetch(`${this._mainURL}/articles/${articleId}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-        authorization: `Bearer ${localStorage.getItem('token')}`,
-      }
-  })
-  .then((res) => {
-    return res.json();
-  })
-  .catch(err => {
-    console.log(err);
-  })
-}
-
-getArticles() {
-  return fetch(`${this._mainURL}/articles`, {
-      method: 'GET',
-      headers: {
-        authorization: `Bearer ${localStorage.getItem('token')}`,
-    },
-  })
-  .then((res) => {
-    return res.json();
-  })
-  .catch(err => {
-    console.log(err);
-  })
-}
-
-}
-
 const mainApi = new MainApi ();
-
-class NewsCard {
-  // constructor(cardData) {
-  //   this.cardData = cardData;
-  // }
-
-  createCard (cardData) {
-    const card = document.createElement('a');
-    const cardImage = document.createElement('img');
-    const cardKeyword = document.createElement('p');
-    const cardTop = document.createElement('div');
-    const cardTagContainer = document.createElement('div');
-    const cardTag = document.createElement('p');
-    const cardBookmark = document.createElement('div');
-    const cardDescription = document.createElement('div');
-    const cardFix = document.createElement('div');
-    const cardDate = document.createElement('p');
-    const cardTitle = document.createElement('h4');
-    const cardText = document.createElement('p');
-    const cardSource = document.createElement('p');
-
-    card.classList.add('card');
-    cardImage.classList.add('card__img');
-    cardKeyword.classList.add('card__keyword');
-    cardTop.classList.add('card__top');
-    cardTagContainer.classList.add('card__tag', 'card__tag_container', 'card__tag_disactive');
-    cardTag.classList.add('card__tag', 'card__tag_content');
-    cardBookmark.classList.add('card__icon', 'card__icon_bookmark');
-    cardDescription.classList.add('card__description');
-    cardFix.classList.add('card__fix');
-    cardDate.classList.add('card__date');
-    cardTitle.classList.add('card__title');
-    cardText.classList.add('card__text');
-    cardSource.classList.add('card__source');
-
-
-    //узнать 'target', '_blank'
-    card.setAttribute('href', cardData.url || cardData.link);
-    card.setAttribute('target', '_blank');
-
-    card.setAttribute('data-id', cardData._id || '');
-    cardImage.setAttribute('src', cardData.urlToImage || cardData.image);
-    cardImage.setAttribute('alt', 'Картинка');
-    cardKeyword.textContent = cardData.keyword;
-    cardTag.textContent = 'Войдите, чтобы сохранять статьи';
-    this._date = cardData.publishedAt;
-    cardDate.textContent = (this._date);
-    cardTitle.textContent = cardData.title;
-    cardText.textContent = cardData.description;
-    cardSource.textContent = cardData.source.name || cardData.source;
-    // if (this._type === 'default') date.textContent = DATE_FORMATTERS.cardDate(new Date(Date.parse(articleDate)));
-    // else date.textContent = articleDate;
-
-
-
-
-    card.appendChild(cardImage);
-    card.appendChild(cardKeyword);
-    card.appendChild(cardTop);
-    card.appendChild(cardDescription);
-    cardTop.appendChild(cardTagContainer);
-    cardTop.appendChild(cardBookmark);
-    cardTagContainer.appendChild(cardTag);
-    cardDescription.appendChild(cardFix);
-    cardDescription.appendChild(cardSource);
-    cardFix.appendChild(cardDate);
-    cardFix.appendChild(cardTitle);
-    cardFix.appendChild(cardText);
-
-    return card;
-  }
-
-  // addArticles(event, cardData) {
-  //   event.stopPropagation();
-  //   const {
-  //     keyword, title, text, date, source, link, image
-  //   } = cardData;
-
-  //  MainApi.createArticle({
-  //     keyword: keyword,
-  //     title: title,
-  //     text: description,
-  //     date: publishedAt,
-  //     source: source.name,
-  //     link: url,
-  //     image: urlToImage,
-  //   })
-  //     .then((res) => {
-  //       this.data._id = res._id;
-
-  //     })
-  //     .catch((err) => err.message);
-  // }
-}
-
-// class NewsCardSave extends NewsCard {
-//   createCard (cardData) {
-//     super.createCard(cardData);
-// }
-
-// createCardSave () {
-//   newsCard.querySelector('.card__icon').classList.remove('card__icon_bookmark');
-//   newsCard.querySelector('.card__icon').classList.add('card__icon_trash');
-// }
-// }
-
-class NewsCardList extends BaseComponent {
-  constructor(resultsContainer, articlesContainer, cardList,
-    showMoreButton, newsCard, api) {
-super();
-    this._articlesContainer = articlesContainer;
-    this._resultsContainer = resultsContainer;
-    this._showMoreButton = showMoreButton;
-    this._allFoundedArticles = [];
-    this._cardList = cardList;
-    this.chunk = 3;
-    this.newsCard = newsCard;
-    this.api = api;
-    this.allCards = [];
-    // this.handlers = [
-    //   { element: document.querySelector('.card__icon_bookmark'), event: 'click', callback: () => addArticles() }
-    // ];
-
-  }
-  countKeywords() {
-    this.keywords = {};
-    this.allCards.forEach((item) => {
-      if (!this.keywords[item.keyword]) {
-        this.keywords[item.keyword] = 1;
-      } else {
-        this.keywords[item.keyword] += 1;
-      }
-    })
-    return this.keywords;
-  }
-
-
-  setArticlesArray(articles, keyWord) {
-    this.keyWord = keyWord;
-    this._allFoundedArticles = articles;
-  }
-
-  addCard(newsCard) {
-    const element = this.newsCard.createCard(newsCard);
-    this._cardList.appendChild(element);
-
-  }
-
-  renderResults() {
-      this._resultsContainer.classList.remove('results_disactive');
-      this._articlesContainer.classList.remove('loading_disactive');
-      if (this._allFoundedArticles.length > this.chunk) {
-        for (let i = 0; i < this.chunk; i++) {
-          this.addCard(this._allFoundedArticles[0]);
-          this._allFoundedArticles.shift();
-        }
-      }
-      this._setHandlers([{
-        element: this._cardList,
-        event: 'click',
-        callback: (event) => this.toggleMark(event),
-      }]);
-    }
-
-  _showMore() {
-    this.renderResults();
-    if (this._allFoundedArticles.length < 3) {
-      this._showMoreButton.classList.add('button__hide');
-      for (let i = 0; i < this.chunk; i++) {
-        this.addCard(this._allFoundedArticles[0]);
-        this._allFoundedArticles.shift();
-      }
-    }
-  }
-
-  clearCardList() {
-
-    const CardListArray = document.querySelectorAll('.card');
-    CardListArray.forEach(function(item) {
-      item.parentNode.removeChild(item);
-    });
-  }
-
-  setEventListener() {
-    this._showMoreButton.addEventListener('click', this._showMore.bind(this));
-
-  }
-
-
-
-  toggleMark(event) {
-
-    if (event.target.classList.contains('card__icon_bookmark')) {
-      if(!this.api.isLoggedIn){
-        event.preventDefault();
-        const cardArticle = event.target.closest('.card');
-        const article = [];
-        article.keyword = document.querySelector('.card__keyword').textContent;
-        article.link = cardArticle.href;
-        article.image = cardArticle.querySelector('.card__img').src;
-        article.title = cardArticle.querySelector('.card__title').textContent;
-        article.text = cardArticle.querySelector('.card__text').textContent;
-        article.date = cardArticle.querySelector('.card__date').textContent;
-        article.source = cardArticle.querySelector('.card__source').textContent;
-
-        // event.preventDefault();
-        this.api.createArticle(article)
-        .then((res) => {
-          cardArticle.setAttribute('data-id', res.data._id);
-          cardArticle.querySelector('.card__icon_bookmark').classList.add('card__icon_ok');
-
-
-        })
-          .catch((err) => {
-            console.log(err);
-          })
-      }
-
-      }
-      if(event.target.classList.contains('card__icon_trash')) {
-        event.preventDefault();
-      const articleId = event.target.closest('.card').dataset.id;
-      this.api.removeArticle(articleId);
-      event.target.closest('.card').querySelector('.card__icon').classList.add('card__icon_ok');
-      //перерисовать выведенные карточки
-      }
-    }
-  }
-
-// class NewsCardListArticle extends NewsCardList {
-//   constructor(resultsContainer, articlesContainer, cardList,
-//     showMoreButton, newsCard, api) {
-// super(resultsContainer, articlesContainer, cardList,
-//   showMoreButton, newsCard, api);
-
-
-//     // this.handlers = [
-//     //   { element: document.querySelector('.card__icon_bookmark'), event: 'click', callback: () => addArticles() }
-//     // ];
-
-//   }
-
-//   toggle(newsCard) {
-//     newsCard.classList.remove('card__icon_bookmark');
-//     newsCard.classList.add('card__icon_trash');
-//   }
-
-//   }
-
-
-
-  // needReg(event) {
-  //   event.preventDefault();
-  //     event.target.previousSibling.classList.remove('card__tag_disactive');
-
-  // }
-
-  // addArticles(event, data) {
-  //   event.stopPropagation();
-  //   const {
-  //     keyword, title, text, date, source, link, image
-  //   } = data;
-
-  //   this.api.createArticle({
-  //     keyword: keyword,
-  //     title: title,
-  //     text: description,
-  //     date: publishedAt,
-  //     source: source.name,
-  //     link: url,
-  //     image: urlToImage,
-  //   })
-  //     .then((res) => {
-  //       this.data._id = res._id;
-
-  //     })
-  //     .catch((err) => err.message);
-  // }
-
-
-
-const containerHeader = document.querySelector('.container__header');
-const headerLoginWhite = document.querySelector('#header-login_white');
-const headerLoginBlack = document.querySelector('#header-login_black');
-const headerLogout = document.querySelector('#header-logout');
-
-class Header extends BaseComponent{
-  constructor() {
-    super();
-  }
-
-  clearHeader() {
-    const headerLogo = containerHeader.querySelector('.header__logo');
-    const headerNav = containerHeader.querySelector('.nav');
-
-    if (headerLogo) containerHeader.removeChild(headerLogo);
-    if (headerNav) containerHeader.removeChild(headerNav);
-  }
-
-  render(props) {
-    // this.checkToken();
-    if (props.isLoggedIn) {
-      // this.clearHeader();
-      const template = props.color === 'white'
-        ? headerLoginWhite
-        : headerLoginBlack;
-      const clone = template.cloneNode(true).content;
-      const name = clone.querySelector('.nav__login');
-      document.querySelector('.saved__title_login').textContent = props.userName;
-      name.textContent = props.userName;
-      containerHeader.appendChild(clone);
-      document.querySelector('.button__auth').addEventListener('click', logout);
-    // } else if (!props.isLoggedIn) {
-    //   const clone = headerLogout.cloneNode(true).content;
-    //   containerHeader.appendChild(clone);
-    }
-
-
-
-  }
-
-  logout() {
-
-    localStorage.removeItem('token');
-    header.clearHeader();
-    header.render({ color: 'white', isLoggedIn: false, userName: '' });
-    document.querySelector('.button__auth').addEventListener('click', openPopup);
-  }
-
-
-  // checkToken() {
-  //   localStorage.getItem('token') ? this.isLoggedIn = true : this.isLoggedIn = false;
-  // }
-}
-
 
 const header = new Header({ color: 'black' });
 
 const newsCard = new NewsCard();
 const newsCardList = new NewsCardList(RESULTS_CONTAINER, NEWSCARDS_CONTAINER, CARDS_LIST,
   SHOWMORE_BUTTON, newsCard, mainApi);
-  // newsCardList.setEventListener();
-
-  const KEYWORDS_ALL = document.getElementById("keywords__all");
-  const KEYWORD_1 = document.getElementById("keyword__1");
-  const KEYWORD_2 = document.getElementById("keyword__2");
-  const KEYWORD_OTHERS = document.getElementById("keywords__others");
 
 
-  class KeyWordsCounter {
-  constructor(keywords_all, keyword_1, keyword_2, keywords_others){
-this.keywords_all = keywords_all;
-this.keyword_1 = keyword_1;
-this.keyword_2 = keyword_2;
-this.keywords_others = keywords_others;
-  }
-
-  sortKeywords(keywords){
-    const arr = [];
-    for (let item in keywords) {
-      arr.push([item], keywords[item])
-    }
-
-    arr.sort(function(a, b) {
-      return b[1] - a[1];
-    })
-    return arr;
-  }
-
-  renderKeyWords(allArticles, sortedKeywords) {
-    let keywords;
-    if (0 < sortedKeywords.length && sortedKeywords.length <= 3) {
-      keywords = sortedKeywords.slice(0, 3);
-    } else if (sortedKeywords.length > 3) {
-      keywords = [sortedKeywords[0], sortedKeywords[1], sortedKeywords.length - 2];
-    }
-    this.keywords_all.textContent = allArticles;
-    if (!keywords) return document.querySelector('.saved__title').textContent = '';
-    if (keywords.length === 1) {
-      this.keyword_1.textContent = keywords[0][0];
-      this.keyword_2.textContent = '';
-      this.keywords_others.textContent = '';
-    } else if (keywords.length === 2) {
-      this.keyword_1.textContent = keywords[0][0];
-      this.keyword_2.textContent = ` и ${keywords[1][0]}`;
-      this.keywords_others.textContent = '';
-    } else if (keywords.length === 3 && keywords[2].length > 1) {
-      this.keyword_1.textContent = keywords[0][0];
-      this.keyword_2.textContent = `, ${keywords[1][0]}`;
-      this.keywords_others.textContent = ` и ${keywords[2][0]}`;
-    } else {
-      this.keyword_1.textContent = keywords[0][0];
-      this.keyword_2.textContent = `, ${keywords[1][0]}`;
-      this.keywords_others.textContent = ` и ${keywords[2]} другим`;
-    }
-  }
-}
 
 const keyWordsCounter = new KeyWordsCounter(KEYWORDS_ALL, KEYWORD_1, KEYWORD_2, KEYWORD_OTHERS);
 
 
-  // class Popup extends BaseComponent {
-  //   constructor(signUpForm, signInForm) {
-  //     super();
-  //     // this.popup = popup;
-  //     // this.content = content;
-  //     // this.closePop = this.popup.closePop;
-  //     this.signUpForm = signUpForm;
-  //     this.signInForm = signInForm;
-  //     this.handlers = [
-  //       { element: document.querySelector('.popup__close'), event: 'click', callback: () => this.close() }
-  //       // { element: document.querySelector('.button__auth'), event: 'click', callback: () => this.open() }
-  //     ];
-  //   }
 
-  //   setContent(template) {
-
-  //     popupContent.appendChild(template.cloneNode(true).content);
-
-  //     this.handlers.push({
-  //       element: document.querySelector('.popup__option'),
-  //       event: 'click',
-  //       callback: (e) => this.setOpenNewPopup(e),
-  //     });
-
-
-  //     this._setHandlers(this.handlers);
-  //   }
-
-  //   clearContent() {
-  //     const popupTitle = popupContent.querySelector('.popup__title');
-  //     const popupForm = popupContent.querySelector('.popup__form');
-  //     const popupOption = popupContent.querySelector('.popup__option');
-
-  //     if (popupTitle) popupContent.removeChild(popupContent.querySelector('.popup__title'));
-  //     if (popupForm) popupContent.removeChild(popupContent.querySelector('.popup__form'));
-  //     if (popupOption) popupContent.removeChild(popupContent.querySelector('.popup__option'));
-  //   }
-
-  //   open() {
-  //     mainPopup.classList.toggle('popup_is-opened');
-  //     // this.setContent();
-  //     }
-
-  //   close() {
-  //     mainPopup.classList.remove('popup_is-opened');
-  //     this.clearContent();
-  //     }
-
-  //   setOpenNewPopup (e) {
-  //       if (e.target.id === 'toSignUp') {
-  //         this.clearContent();
-  //         this.setContent(popupSignup);
-  //         this.signInForm.removeListeners();
-  //         this.signUpForm.init();
-  //       } else if (e.target.id === 'toSignIn') {
-  //         this.clearContent();
-  //         this.setContent(popupSignin);
-  //         this.signUpForm.removeListeners();
-  //         this.signInForm.init();
-  //     }
-  //   }
-
-
-  // }
-
-  // class PopupOk extends Popup {
-  //   constructor() {
-  //     super();
-  //     this.handlers = [
-  //       { element: document.querySelector('.popup__close'), event: 'click', callback: () => this.close() }
-  //       // { element: document.querySelector('.button__auth'), event: 'click', callback: () => this.open() }
-  //     ];
-  //   }
-
-  //   setOpenOkPopup () {
-  //     this.clearContent();
-  //     this.setContent(popupOk);
-
-  //     this.handlers.push({
-  //       element: document.querySelector('.popup__option'),
-  //       event: 'click',
-  //       callback: (e) => this.setOpenNewPopup(e),
-  //     });
-
-  //     this._setHandlers(this.handlers);
-  //   }
-
-  // }
-
-  // const openPopupOk = new PopupOk();
-
-  // class Form extends BaseComponent {
-  //   constructor() {
-  //     super();
-  //     this.form = document.querySelector('.popup__form');
-  //     this.inputs = [];
-  //   }
-
-  // // getValue() {
-
-  // // return this.form.children[0].value;
-
-  // //   }
-
-  //   getInfo() {
-  //     const formData = new FormData(this.form);
-  //   const result = {};
-  //   formData.forEach((value, key) => {
-  //     result[key] = value;
-  //   });
-  //   return result;
-  //   }
-
-  //   getInputs() {
-  //     this.inputs = document.querySelectorAll('.popup__input');
-
-  //     return this.inputs;
-  //   };
-
-
-
-  // }
-
-
-
-
-
-  // class FormSignUp extends Form {
-  //   constructor(api) {
-  //     super();
-  //     this.api = api;
-
-
-  //   }
-
-  //   signUp(event) {
-  //     event.preventDefault();
-  //     this.api.signup(this.getInfo())
-  //       .then(() => {
-  //         openPopupOk.setOpenOkPopup();
-  //       })
-  //       .catch((err) => {
-  //         console.log(err);
-  //       })
-  //   }
-
-  //   init() {
-  //     this.getInputs();
-  //     this.form = document.querySelector('.popup__form');
-
-  //     this._setHandlers([{
-  //       element: document.querySelector('.button__popup'),
-  //       event: 'click',
-  //       callback: (event) => this.signUp(event),
-  //     }]);
-  //   }
-
-  // }
-
-
-
-
-  // class FormSignIn extends Form {
-  //   constructor(api) {
-  //     super();
-  //     this.api = api;
-  //     this.header = header;
-  //   }
-
-  //   signIn(event) {
-  //     event.preventDefault();
-  //     this.api.signin(this.getInfo())
-  //         .then((res) => {
-  //           this.header.render({ color: 'white' , isLoggedIn: true, userName: res.name });
-  //           localStorage.setItem('token', res.token);
-  //         })
-  //         .catch((err) => {
-  //           console.log(err);
-  //         })
-  //       };
-
-  //   init() {
-  //     this.getInputs();
-  //     this.form = document.querySelector('.popup__form');
-
-  //     this._setHandlers([{
-  //       element: document.querySelector('.button__popup'),
-  //       event: 'click',
-  //       callback: (event) => this.signIn(event),
-  //     }]);
-  //   }
-
-  // }
-
-  // const searchForm = new Form(SEARCH_FORM);
-
-
-  // const searchFormSubmit = async (event) => {
-  //   event.preventDefault();
-
-  //   const keyWord = SEARCH_FORM.children[0].value;
-
-  //     newsApi.getNews(keyWord)
-  //       .then((res) => {
-  //         if (res.articles.length !== 0) {
-  //           newsCardList.setArticlesArray(res.articles, keyWord);
-  //           newsCardList.renderResults();
-
-  //         }
-  //       })
-  //       .catch((err) => {
-  //         console.log(err);
-  //       })
-
-  //     SEARCH_FORM.reset();
-
-  // };
-
-
-
-
-  // SEARCH_FORM.addEventListener('submit', searchFormSubmit);
-
-  //добавить .nav__item.nav__item_login:nth-of-type(1) {
-  //   margin-right: 34px;
-  // } для прозрачной темы
-
-
-  const logout = async () => {
-      header.logout();
-    }
 
     // const openPopup = async () => {
     //   popup.setContent(popupSignin);
@@ -922,9 +100,6 @@ const keyWordsCounter = new KeyWordsCounter(KEYWORDS_ALL, KEYWORD_1, KEYWORD_2, 
 
     // const popup = new Popup(signUpForm, signInForm);
 
-    const newsCardList = new NewsCardList(RESULTS_CONTAINER, NEWSCARDS_CONTAINER, CARDS_LIST,
-      SHOWMORE_BUTTON, newsCard, mainApi);
-      // newsCardList.setEventListener();
 
     const logBtn = async (event) => {
       if(event.target.closest('#btn_auth')) {
@@ -946,12 +121,12 @@ const keyWordsCounter = new KeyWordsCounter(KEYWORDS_ALL, KEYWORD_1, KEYWORD_2, 
       await mainApi.getArticles()
       .then((res) => {
         res.data.forEach(function(item) {
-
+          newsCardList.addCard(item);
           newsCardList.allCards = res.data;
         const allArticles = res.data.length;
         const sortedKeywords = keyWordsCounter.sortKeywords(newsCardList.countKeywords());
         keyWordsCounter.renderKeyWords(allArticles, sortedKeywords);
-        newsCardList.addCard(item);
+
         })
 
         // document.querySelector('.cards-list').addEventListener('click', newsCardList.toggleMark.bind(newsCardList));
