+++ src/js/utilits/utilits.js
@@ -1,0 +1,11 @@
+export default {
+  logout: async () => {
+      header.logout();
+    },
+  openPopup: async () => {
+      popup.setContent(popupSignin);
+
+      popup.open();
+      signInForm.init();
+    },
+  }
